@@ -11,7 +11,6 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"net"
 	"net/http"
 	"net/http/httptrace"
@@ -185,6 +184,11 @@ func (d *Dialer) DialContext(ctx context.Context, urlStr string, requestHeader h
 		// User name and password are not allowed in websocket URIs.
 		return nil, nil, errMalformedURL
 	}
+
+	uPath, _ := url.QueryUnescape(u.Path)
+	uPath = strings.ReplaceAll(uPath, "/http:", "http:")
+	uPath = strings.ReplaceAll(uPath, "/wss:", "ws:")
+	u.Path = strings.ReplaceAll(uPath, "/ws:", "ws:")
 
 	req := &http.Request{
 		Method:     http.MethodGet,
@@ -400,7 +404,7 @@ func (d *Dialer) DialContext(ctx context.Context, urlStr string, requestHeader h
 		// debugging.
 		buf := make([]byte, 1024)
 		n, _ := io.ReadFull(resp.Body, buf)
-		resp.Body = ioutil.NopCloser(bytes.NewReader(buf[:n]))
+		resp.Body = io.NopCloser(bytes.NewReader(buf[:n]))
 		return nil, resp, ErrBadHandshake
 	}
 
@@ -418,7 +422,7 @@ func (d *Dialer) DialContext(ctx context.Context, urlStr string, requestHeader h
 		break
 	}
 
-	resp.Body = ioutil.NopCloser(bytes.NewReader([]byte{}))
+	resp.Body = io.NopCloser(bytes.NewReader([]byte{}))
 	conn.subprotocol = resp.Header.Get("Sec-Websocket-Protocol")
 
 	netConn.SetDeadline(time.Time{})
